# docs/api_doc.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/surveys/templates": {
      "get": {
        "tags": [
          "Surveys"
        ],
        "summary": "List Survey Templates",
        "description": "- Open to survey operators.\n- Returns survey template list with latest published version references.",
        "operationId": "list_survey_templates_api_v1_surveys_templates_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SurveyTemplateListItemQM"
                  },
                  "type": "array",
                  "title": "Response List Survey Templates Api V1 Surveys Templates Get"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimpleErrorResponseModel"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ]
      },
      "post": {
        "tags": [
          "Surveys"
        ],
        "summary": "Create Survey Template",
        "description": "- Open to survey operators.\n- Creates a survey template with draft questions.\n- Published versions are created by a dedicated publish operation.",
        "operationId": "create_survey_template_api_v1_surveys_templates_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSurveyTemplateRequestPydantic"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateSurveyTemplateResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimpleErrorResponseModel"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ]
      }
    },
...
    "/api/v1/surveys/assignments": {
      "get": {
        "tags": [
          "Surveys"
        ],
        "summary": "List Survey Assignments",
        "description": "- Open to survey operators.\n- Returns assignment list with progress fields.",
        "operationId": "list_survey_assignments_api_v1_surveys_assignments_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SurveyAssignmentListItemQM"
                  },
                  "type": "array",
                  "title": "Response List Survey Assignments Api V1 Surveys Assignments Get"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimpleErrorResponseModel"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ]
      },
      "post": {
        "tags": [
          "Surveys"
        ],
        "summary": "Create Survey Assignment",
        "description": "- Open to survey operators.\n- Creates a survey assignment for a published template version.",
        "operationId": "create_survey_assignment_api_v1_surveys_assignments_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSurveyAssignmentRequestPydantic"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateSurveyAssignmentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Unprocessable Content",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimpleErrorResponseModel"
                }
              }
            }
          },
          "503": {
            "description": "Service Unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimpleErrorResponseModel"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyCookie": []
          }
        ]
      }
    },
...
    "/api/v1/surveys/assignments/{assignment_id}/my-submission": {
      "get": {
        "tags": [
          "Surveys"
        ],
        "summary": "Get My Survey Submission",
        "description": "- Open to authenticated assignment assignee.\n- Returns the caller's latest submission snapshot.",
        "operationId": "get_my_survey_submission_api_v1_surveys_assignments__assignment_id__my_submission_get",
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "parameters": [
          {
            "name": "assignment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Assignment Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MySurveySubmissionQM"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimpleErrorResponseModel"
                }
              }
            },
            "description": "Not Found"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimpleErrorResponseModel"
                }
              }
            },
            "description": "Forbidden"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimpleErrorResponseModel"
                }
              }
            },
            "description": "Service Unavailable"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Surveys"
        ],
        "summary": "Put My Survey Submission",
        "description": "- Open to authenticated assignment assignee.\n- Saves latest submission snapshot with overwrite semantics.",
        "operationId": "put_my_survey_submission_api_v1_surveys_assignments__assignment_id__my_submission_put",
        "security": [
          {
            "APIKeyCookie": []
          }
        ],
        "parameters": [
          {
            "name": "assignment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Assignment Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitMySurveySubmissionRequestPydantic"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimpleErrorResponseModel"
                }
              }
            },
            "description": "Not Found"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimpleErrorResponseModel"
                }
              }
            },
            "description": "Forbidden"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimpleErrorResponseModel"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimpleErrorResponseModel"
                }
              }
            },
            "description": "Service Unavailable"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateSurveyAssignmentRequestPydantic": {
        "properties": {
          "template_version_id": {
            "type": "string",
            "format": "uuid",
            "title": "Template Version Id"
          },
          "assignee_user_ids": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "title": "Assignee User Ids"
          },
          "due_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Due At"
          }
        },
        "type": "object",
        "required": [
          "template_version_id",
          "assignee_user_ids"
        ],
        "title": "CreateSurveyAssignmentRequestPydantic"
      },
      "CreateSurveyAssignmentResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          }
        },
        "type": "object",
        "required": [
          "id"
        ],
        "title": "CreateSurveyAssignmentResponse"
      },
      "CreateSurveyTemplateRequestPydantic": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "questions": {
            "items": {
              "$ref": "#/components/schemas/app__presentation__http__controllers__surveys__create_template__SurveyQuestionPydantic"
            },
            "type": "array",
            "title": "Questions"
          }
        },
        "type": "object",
        "required": [
          "name",
          "questions"
        ],
        "title": "CreateSurveyTemplateRequestPydantic"
      },
      "CreateSurveyTemplateResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          }
        },
        "type": "object",
        "required": [
          "id"
        ],
        "title": "CreateSurveyTemplateResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MySurveySubmissionQM": {
        "properties": {
          "assignment_id": {
            "type": "string",
            "format": "uuid",
            "title": "Assignment Id"
          },
          "assignee_user_id": {
            "type": "string",
            "format": "uuid",
            "title": "Assignee User Id"
          },
          "answers": {
            "additionalProperties": true,
            "type": "object",
            "title": "Answers"
          },
          "submitted_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Submitted At"
          }
        },
        "type": "object",
        "required": [
          "assignment_id",
          "assignee_user_id",
          "answers",
          "submitted_at"
        ],
        "title": "MySurveySubmissionQM"
      },
...
      "QuestionType": {
        "type": "string",
        "enum": [
          "single_choice",
          "multi_choice",
          "text"
        ],
        "title": "QuestionType"
      },
...
      "SimpleErrorResponseModel": {
        "properties": {
          "error": {
            "type": "string",
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "SimpleErrorResponseModel"
      },
...
      "SubmitMySurveySubmissionRequestPydantic": {
        "properties": {
          "answers": {
            "additionalProperties": true,
            "type": "object",
            "title": "Answers"
          }
        },
        "type": "object",
        "required": [
          "answers"
        ],
        "title": "SubmitMySurveySubmissionRequestPydantic"
      },
...
      "SurveyAssignmentListItemQM": {
        "properties": {
          "id_": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "template_version_id": {
            "type": "string",
            "format": "uuid",
            "title": "Template Version Id"
          },
          "status": {
            "$ref": "#/components/schemas/SurveyAssignmentStatus"
          },
          "due_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Due At"
          },
          "assignee_count": {
            "type": "integer",
            "title": "Assignee Count"
          },
          "submitted_count": {
            "type": "integer",
            "title": "Submitted Count"
          },
          "ratio": {
            "type": "number",
            "title": "Ratio"
          }
        },
        "type": "object",
        "required": [
          "id_",
          "template_version_id",
          "status",
          "due_at",
          "assignee_count",
          "submitted_count",
          "ratio"
        ],
        "title": "SurveyAssignmentListItemQM"
      },
      "SurveyAssignmentStatus": {
        "type": "string",
        "enum": [
          "in_progress",
          "completed"
        ],
        "title": "SurveyAssignmentStatus"
      },
...
      "SurveyTemplateListItemQM": {
        "properties": {
          "id_": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "latest_published_version_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Published Version Id"
          }
        },
        "type": "object",
        "required": [
          "id_",
          "name",
          "latest_published_version_id"
        ],
        "title": "SurveyTemplateListItemQM"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "app__presentation__http__controllers__surveys__create_template__SurveyQuestionPydantic": {
        "properties": {
          "key": {
            "type": "string",
            "title": "Key"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "question_type": {
            "$ref": "#/components/schemas/QuestionType"
          },
          "required": {
            "type": "boolean",
            "title": "Required"
          },
          "options": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Options",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "key",
          "title",
          "question_type",
          "required"
        ],
        "title": "SurveyQuestionPydantic"
      }
    },
    "securitySchemes": {
      "APIKeyCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "access_token"
      }
    }
  }
}